Fixing integer division resulting in 0 failure rate, and updating documentation on what is included in JobsMetrics.

+++ src/services/a-rex/grid-manager/log/JobsMetrics.cpp
@@ -56,15 +56,10 @@ void JobsMetrics::SetGmetricPath(const char* path) {
   std::string job_id = i->job_id;
 
   /*
-    ## - staging -- number of tasks in different data staging states
-    ## - cache -- free cache space
-    ## - session -- free session directory space
-    ## - heartbeat -- last modification time of A-REX heartbeat
     ## - processingjobs -- the number of jobs currently being processed by ARC (jobs
-    ##                     between PREPARING and FINISHING states)
+    ##                     between PREPARING and FINISHING states) -- TO-DO
     ## - failedjobs -- the ratio of failed jobs to all jobs
     ## - jobstates -- number of jobs in different A-REX internal stages
-    ## - all -- all of the above metrics
   */
   
 
@@ -75,7 +70,7 @@ void JobsMetrics::SetGmetricPath(const char* path) {
 
 
   //ratio of failed jobs to all jobs
-  fail_ratio = job_fail_counter/job_counter;
+  fail_ratio = (double)job_fail_counter/(double)job_counter;
   fail_ratio_changed = true;
 
   //actual states (jobstates)
